p4_lib: revise the P4 libs and the P4 lib parser class
This commit revises all P4 libs, and the P4 lib parser class.
All changes:
1. nfcp_lib_parser -> lemur_p4lib_parser
2. P4 NF modules under p4_14_lib/ are cleaned
