+++ src/p4_14_lib/acl.lib
@@ -1,14 +1,15 @@
-############################################################
-# ACL.LIB
-# This will perform the basic access control list (ACL) functions.
-# You can match MAC, IP, TCP fields easily.
-# You can also define your own header and implement the corresponding
-# actions to get what you want.
 #
-############################################################
+# Title: acl.lib
+# This P4 module implements the basic access control list (ACL) function.
+# It matches on Ethernet, IP, TCP fields. You can also define customized
+# headers and implement corresponding actions to serve your goal.
+#
 
 default_prefix = acl
 
+/*
+ * P4 module marcos
+ */
 add macro(ACL_TABLE_SIZE, 50)
 add macro(TYPE_IPV4, 0x0800)
 add macro(TYPE_TCP, 0x06)
@@ -17,7 +18,6 @@ add macro(TYPE_UDP, 0x11)
 /*
  * Header Definition Part
  */
-
 #define Ethernet
 #define IPv4
 #define TCP
@@ -26,14 +26,11 @@ add macro(TYPE_UDP, 0x11)
 /* Metadata Definition Part
  *
  */
-
 add metadata(acl_table_miss_flag, 1)
-
 
 /*
  * P4 Header Parser Specification Part
  */
-
 set Ethernet(etherType, 1, {TYPE_IPV4:IPv4})
 set IPv4(protocol, 2, {TYPE_TCP:TCP, TYPE_UDP:UDP})
 set TCP(None, 0, {})
@@ -88,9 +85,7 @@ apply {
 
 
 /*
- * P4 Header Deparser Specification Part (Done)
+ * P4 Header Deparser Specification Part
  */
-
 deparser_seq = [IPv4, Ethernet]
 add deparser( deparser_seq )
-
