+++ src/p4_14_lib/pfilter.lib
@@ -1,14 +1,15 @@
-############################################################
-# PFILTER.LIB
-# This is the packet filter module implementation for P4 switch.
-# You can match MAC, IP, TCP fields easily.
-# You can also define your own header and implement the corresponding
-# actions to get what you want.
 #
-############################################################
+# Title: pfilter.lib
+# This P4 module implements a packet filter module. The module can match
+# on Ethernet, IP, TCP fields. You can also define your own header and
+# implement the corresponding actions to filter out other fields.
+#
 
 default_prefix = pfilter
 
+/*
+ * P4 module marcos
+ */
 add macro(PFILTER_TABLE_SIZE, 50)
 add macro(TYPE_IPV4, 0x0800)
 add macro(TYPE_TCP, 0x06)
@@ -17,7 +18,6 @@ add macro(TYPE_UDP, 0x11)
 /*
  * Header Definition Part
  */
-
 #define Ethernet
 #define IPv4
 #define TCP
@@ -26,19 +26,18 @@ add macro(TYPE_UDP, 0x11)
 /* Metadata Definition Part
  *
  */
-
 add metadata(gate_select, 4)
 
 
 /*
  * P4 Header Parser Specification Part
  */
-
 set Ethernet(etherType, 1, {TYPE_IPV4:IPv4})
 set IPv4(protocol, 2, {TYPE_TCP:TCP, TYPE_UDP:UDP})
 set TCP(None, 0, {})
 set UDP(None, 0, {})
 add parser( Ethernet )
+
 
 # MyIngress
 
@@ -73,9 +72,7 @@ apply {
 
 
 /*
- * P4 Header Deparser Specification Part (Done)
+ * P4 Header Deparser Specification Part
  */
-
 deparser_seq = [IPv4, Ethernet]
 add deparser( deparser_seq )
-
