+++ src/p4_14_lib/update_ttl_p4.lib
@@ -1,10 +1,13 @@
-############################################################
-# UPDATE_TTL.LIB
-# This module will decrease the ttl field in the IPv4 header by 1.
 #
-############################################################
+# update_ttl.lib
+# This is a dummy module that simply decreases the IPv4.ttl field by 1.
+#
 
 default_prefix = updatettl
+
+/*
+ * P4 module marcos
+ */
 add macro(TYPE_IPV4, 0x0800)
 
 /*
@@ -54,6 +57,5 @@ apply {
 /*
  * P4 Header Deparser Specification Part
  */
-
 deparser_seq = [IPv4, Ethernet]
 add deparser( deparser_seq )
