+++ src/p4_14_lib/vlan_rm.lib
@@ -1,21 +1,25 @@
-############################################################
-# VLANRM.LIB
-# This will remove vlan header between Ethernet and IP layer.
-# VLAN is used to partition and isolate network at data link layer.
-# Original VLAN header will be discarded.
-# EtherType field of Ethernet header will point to next protocol.
 #
-############################################################
+# Title: vlan_rm.lib
+# This P4 module removes a vlan header for incoming packets if they
+# have one.
+# Note: the VLAN header is commonly used to divide and isolate a
+# network at the link layer.
+# In this implementation, the packet's original VLAN header should
+# be discarded. Ethernet.Ethertype points to the next available
+# protocol (mostly ipv4).
+#
 
 default_prefix = vlan_rm
 
+/*
+ * P4 module marcos
+ */
 add macro(ETHERTYPE_IPV4, 0x0800)
 add macro(ETHERTYPE_VLAN, 0x8100)
 
 /*
  * Header Definition Part
  */
-
 #define Ethernet
 #define VLAN
 #define IPv4
@@ -23,7 +27,6 @@ add macro(ETHERTYPE_VLAN, 0x8100)
 /* Metadata Definition Part
  *
  */
-
 
 /*
  * P4 Header Parser Specification Part
@@ -33,8 +36,8 @@ set VLAN(nextType, 1, {ETHERTYPE_IPV4: IPv4})
 set IPv4(None, 0, {})
 add parser( Ethernet )
 
-# MyIngress
 
+# MyIngress
 
 action do_VLAN_decap() {
     modify_field(ethernet.etherType, vlan.nextType);
@@ -68,11 +71,7 @@ apply {
 
 
 /*
- * P4 Header Deparser Specification Part (Done)
+ * P4 Header Deparser Specification Part
  */
-
 deparser_seq = [IPv4, VLAN, Ethernet]
 add deparser( deparser_seq )
-
-
-
